Add keyboard support to toggle summary row
